moved reducer location
