+++ client/src/app/store.js
@@ -1,7 +1,7 @@
 import { createStore, applyMiddleware } from "redux";
 import { composeWithDevTools } from "redux-devtools-extension";
 import thunk from "redux-thunk";
-import rootReducer from "../old-ui-src/reducers";
+import rootReducer from "../reducers";
 
 const initialState = {};
